Reshuffle remaining files.
As far as OSX/Windows is concerned 'Da' = 'DA' so all of the minor modules are
removed and condensed into packages for lower, upper, and mixed case,
respectively.

unsure if this is the best way to go...
